a2alloc.c: use smaller sizes
storing the size of free or used memory can be limitted to a uint32_t
because 2^32>256MB, which the max number of bytes that our global memory
can be. don't use size_t is dependent on the arch iirc, which poses a slight
problem (maybe), as it's only guaranteed to be 16B, but I think it might
go up to 64B.

reduce the size of the free flag. a bool is 1B, which is nice and small,
rather than an int which is larger.